+++ webviz_config/utils/__init__.py
@@ -2,3 +2,6 @@ from ._localhost_open_browser import LocalhostOpenBrowser
 from ._available_port import get_available_port
 from ._silence_flask_startup import silence_flask_startup
 from ._dash_component_utils import calculate_slider_step
+from ._deprecate_webviz_settings_attribute_in_dash_app import (
+    deprecate_webviz_settings_attribute_in_dash_app,
+)